Add contest scaffolding and validation to backend
Added the contest scaffolding so users(clients) can create contests.
Contests inherit from protected controller so that you can't navigate
to them on the front end unless you are signed in. This should also
limit access in the backend. I still need to address the issue where All
Logged in users can see all contests. I believe this is a back end issue
but there may be things I need to add to the front end as well.